+++ src/bobbin/interaction/Printers.java
@@ -47,8 +47,8 @@ public class Printers {
      * @param writer     to print prompt to
      * @param collection for the player to choose from
      */
-    public static void printUnordered(PrintWriter writer, Collection<BaseInteractive> collection) {
-        for (BaseInteractive gameEntity : collection) {
+    public static void printUnordered(PrintWriter writer, Collection<Interactive> collection) {
+        for (Interactive gameEntity : collection) {
             writer.println(format("Prompts.unorderedList", gameEntity));
 
             gameEntity.setSeen();
@@ -62,9 +62,9 @@ public class Printers {
      * @param writer to print prompt to
      * @param list   for the player to choose from
      */
-    public static <T extends BaseInteractive> void printOrdered(PrintWriter writer, List<T> list) {
+    public static <T extends Interactive> void printOrdered(PrintWriter writer, List<T> list) {
         int i = 1;
-        for (BaseInteractive gameEntity : list) {
+        for (Interactive gameEntity : list) {
             writer.println(format("Prompts.orderedList", i, gameEntity));
 
             ++i;
@@ -79,11 +79,11 @@ public class Printers {
      * @param writer to print prompt to
      * @param array  for the player to choose from
      */
-    public static void printOrdered(PrintWriter writer, BaseInteractive[] array) {
+    public static void printOrdered(PrintWriter writer, Interactive[] array) {
         printOrdered(writer, Arrays.asList(array));
     }
 
-    public static void print(PrintWriter writer, BaseInteractive gameEntity) {
+    public static void print(PrintWriter writer, Interactive gameEntity) {
         writer.print(Globals.messages.getString("Messages.prefix.primary"));
         writer.println(gameEntity.toString());
         writer.flush();
