+++ src/bobbin/interaction/actions/ActionList.java
@@ -5,12 +5,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import bobbin.characters.PlayerCharacter;
-import bobbin.interaction.BaseInteractive;
+import bobbin.interaction.Interactive;
 
 /**
  * Problems this is solving:
  * - Easily accessible, consistent list of actions to be taken
- * - Keeps {@link BaseInteractive}s fresh, as they are fetched only when the player is prompted
+ * - Keeps {@link Interactive}s fresh, as they are fetched only when the player is prompted
  * - More extensible than other options. Can keep Actions as a group of constants
  */
 public class ActionList extends ArrayList<BaseAction> {
@@ -18,9 +18,9 @@ public class ActionList extends ArrayList<BaseAction> {
     /**
      * Apply all of the {@link Action}s in the {@link ActionList}.
      *
-     * @return {@link List} of {@link BaseInteractive} objects returned by {@link Action}s.
+     * @return {@link List} of {@link Interactive} objects returned by {@link Action}s.
      */
-    public List<BaseInteractive> build(PlayerCharacter playerCharacter) {
+    public List<Interactive> build(PlayerCharacter playerCharacter) {
         return stream().map((action -> action.apply(playerCharacter))).collect(Collectors.toList());
     }
 }
