+++ test/bobbin/usability/BaseUsabilityTest.java
@@ -3,19 +3,19 @@ package bobbin.usability;
 import java.io.BufferedReader;
 
 import bobbin.interaction.ExitToException;
-import bobbin.interaction.BaseInteractive;
+import bobbin.interaction.Interactive;
 import bobbin.unit.interaction.BaseConsoleTest;
 
 public class BaseUsabilityTest extends BaseConsoleTest {
 
     protected void run(BufferedReader reader) {
-        BaseInteractive next = playerCharacter;
+        Interactive next = playerCharacter;
 
         while (true) {
             try {
                 next.interact(playerCharacter, null, reader, writer);
             }
-            catch (BaseInteractive.ResetStackException e) {
+            catch (Interactive.ResetStackException e) {
                 next = e.then;
             }
             catch (ExitToException e) {
